requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
aiohttp==3.9.1
flask==3.0.0
flask-cors==4.0.0 